convert TabbedPanel to ts
